Move SwapLayout to it's own namespace
