--- src/components/SwapLayout/Card/index.tsx
@@ -1,1 +1,0 @@
-export * from 'components/SwapLayout/Card/Card'
