+++ src/pages/Trades.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import TradesPage from 'components/TradesWidget'
-import { PageWrapper } from 'components/SwapLayout/PageWrapper'
+import { PageWrapper } from 'components/layout/SwapLayout/PageWrapper'
 
 const Trades: React.FC = () => (
   <PageWrapper>
